+++ src/imgvisfeat/GradientVisualizer.py
@@ -24,22 +24,22 @@ class ColorGradientVisualizer(AbstractVisualizer):
         """
         image = source.copy()
         if image.ndim != 3:
-            image = cv2.cvtColor(image, cv2.COLOR_GRAY2BGR)  # type: ignore
+            image = cv2.cvtColor(image, cv2.COLOR_GRAY2BGR).astype(np.uint8)
         grad_x = np.zeros_like(image, dtype=np.uint8)
         grad_y = np.zeros_like(image, dtype=np.uint8)
         grad_xy = np.zeros_like(image, dtype=np.uint8)
         kernel_x = np.array([[0, 0, 0], [-1, 0, 1], [0, 0, 0]], np.float32)
         kernel_y = np.array([[0, 1, 0], [0, 0, 0], [0, -1, 0]], np.float32)
         kernel_xy = np.array([[0, 1, 0], [-1, 0, 1], [0, -1, 0]], np.float32)
-        grad_x[:, :, 0] = cv2.filter2D(image[:, :, 0], -1, kernel_x)
-        grad_y[:, :, 0] = cv2.filter2D(image[:, :, 0], -1, kernel_y)
-        grad_xy[:, :, 0] = cv2.filter2D(image[:, :, 0], -1, kernel_xy)
-        grad_x[:, :, 1] = cv2.filter2D(image[:, :, 1], -1, kernel_x)
-        grad_y[:, :, 1] = cv2.filter2D(image[:, :, 1], -1, kernel_y)
-        grad_xy[:, :, 1] = cv2.filter2D(image[:, :, 1], -1, kernel_xy)
-        grad_x[:, :, 2] = cv2.filter2D(image[:, :, 2], -1, kernel_x)
-        grad_y[:, :, 2] = cv2.filter2D(image[:, :, 2], -1, kernel_y)
-        grad_xy[:, :, 2] = cv2.filter2D(image[:, :, 2], -1, kernel_xy)
+        grad_x[:, :, 0] = cv2.filter2D(image[:, :, 0], -1, kernel_x).astype(np.uint8)
+        grad_y[:, :, 0] = cv2.filter2D(image[:, :, 0], -1, kernel_y).astype(np.uint8)
+        grad_xy[:, :, 0] = cv2.filter2D(image[:, :, 0], -1, kernel_xy).astype(np.uint8)
+        grad_x[:, :, 1] = cv2.filter2D(image[:, :, 1], -1, kernel_x).astype(np.uint8)
+        grad_y[:, :, 1] = cv2.filter2D(image[:, :, 1], -1, kernel_y).astype(np.uint8)
+        grad_xy[:, :, 1] = cv2.filter2D(image[:, :, 1], -1, kernel_xy).astype(np.uint8)
+        grad_x[:, :, 2] = cv2.filter2D(image[:, :, 2], -1, kernel_x).astype(np.uint8)
+        grad_y[:, :, 2] = cv2.filter2D(image[:, :, 2], -1, kernel_y).astype(np.uint8)
+        grad_xy[:, :, 2] = cv2.filter2D(image[:, :, 2], -1, kernel_xy).astype(np.uint8)
         return GradientResult(grad_x, grad_y, grad_xy)
 
 
@@ -60,7 +60,7 @@ class GrayGradientVisualizer(AbstractVisualizer):
             GradientResult: The gradient in the x, y, and xy directions.
         """
         if source.ndim != 2:
-            gray = cv2.cvtColor(source, cv2.COLOR_BGR2GRAY)
+            gray = cv2.cvtColor(source, cv2.COLOR_BGR2GRAY).astype(np.uint8)
         else:
             gray = source.copy()
         kernel_x = np.array([[0, 0, 0], [-1, 0, 1], [0, 0, 0]], np.float32)
